requirements.txt
dash==2.11.0
dash-bootstrap-components==1.5.0
flask>=1.0.4,<2.3.0
gunicorn
opencv-python-headless
numpy
pandas
xlsxwriter
mediapipe==0.10.21
# ← NUEVAS PARA EL INFORME DOCX
python-docx
Pillow            # la usa python-docx para manejar imágenes